Ignore BIND_ABOVE_CLIENT for same-process connections
Binding to a service using BIND_ABOVE_CLIENT affects the OOM adjustment
of both the client and the service; client's value is dropped by one
level in modifyRawOomAdj() to make sure it's lower than the service's.

Doing this unconditionally, however, means that the process' OOM score
will be reduced if binding to a service that lives in the same process
as the client. For example, a top app would have its oom_score_adj set
to 100 (visible) rather than 0 (foreground).

Test: atest MockingOomAdjusterTests
Signed-off-by: Tomislav Novak <tnovak@meta.com>
Change-Id: Iee5024b8f13771bc98a39a20e5f96a89a4c79b4e